+++ src/session/src/lib.rs
@@ -2,7 +2,19 @@
 // Licensed under the Apache License, Version 2.0
 // http://www.apache.org/licenses/LICENSE-2.0
 
+//! Abstractions for bi-directional buffered communications on top of streams.
+//! This allows for efficient reading and writing for stream-oriented
+//! communication and provides abstractions for request/response oriented
+//! client/server communications.
+
 pub use buffer::*;
+
+mod client;
+mod server;
+
+pub use client::ClientSession;
+pub use server::ServerSession;
+
 use std::os::unix::prelude::AsRawFd;
 
 use ::net::*;
@@ -276,400 +288,5 @@ impl event::Source for Session {
 
     fn deregister(&mut self, registry: &Registry) -> Result<()> {
         self.stream.deregister(registry)
-    }
-}
-
-/// A basic session to represent the client side of a framed session.
-pub struct ClientSession<Parser, Tx, Rx> {
-    session: Session,
-    parser: Parser,
-    pending: VecDeque<(Instant, Tx)>,
-    _rx: PhantomData<Rx>,
-}
-
-impl<Parser, Tx, Rx> Debug for ClientSession<Parser, Tx, Rx> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::result::Result<(), std::fmt::Error> {
-        write!(f, "{:?}", self.session)
-    }
-}
-
-impl<Parser, Tx, Rx> AsRawFd for ClientSession<Parser, Tx, Rx> {
-    fn as_raw_fd(&self) -> i32 {
-        self.session.as_raw_fd()
-    }
-}
-
-impl<Parser, Tx, Rx> ClientSession<Parser, Tx, Rx>
-where
-    Tx: Compose,
-    Parser: Parse<Rx>,
-{
-    pub fn new(session: Session, parser: Parser) -> Self {
-        Self {
-            session,
-            parser,
-            pending: VecDeque::with_capacity(256),
-            _rx: PhantomData,
-        }
-    }
-
-    /// Sends the frame to the underlying session but does *not* flush the
-    /// session buffer. This function also adds a timestamp to a queue so that
-    /// response latencies can be determined. The latency will include any time
-    /// that it takes to compose the message onto the session buffer, time to
-    /// flush the session buffer, and any additional calls to flush which may be
-    /// required.
-    pub fn send(&mut self, tx: Tx) -> Result<usize> {
-        CLIENT_SESSION_SEND.increment();
-        let now = Instant::now();
-        let size = tx.compose(&mut self.session);
-        self.pending.push_back((now, tx));
-        Ok(size)
-    }
-
-    pub fn receive(&mut self) -> Result<(Tx, Rx)> {
-        let src: &[u8] = self.session.borrow();
-        match self.parser.parse(src) {
-            Ok(res) => {
-                CLIENT_SESSION_RECV.increment();
-                let now = Instant::now();
-                let (timestamp, request) = self
-                    .pending
-                    .pop_front()
-                    .ok_or_else(|| Error::from(ErrorKind::InvalidInput))?;
-                let latency = now - timestamp;
-                CLIENT_RESPONSE_LATENCY.increment(now, latency.as_nanos(), 1);
-                let consumed = res.consumed();
-                let msg = res.into_inner();
-                self.session.consume(consumed);
-                Ok((request, msg))
-            }
-            Err(e) => {
-                if e.kind() != ErrorKind::WouldBlock {
-                    CLIENT_SESSION_RECV_EX.increment();
-                }
-                Err(e)
-            }
-        }
-    }
-
-    pub fn flush(&mut self) -> Result<()> {
-        self.session.flush()?;
-        Ok(())
-    }
-
-    pub fn write_pending(&self) -> usize {
-        self.session.write_pending()
-    }
-
-    pub fn fill(&mut self) -> Result<usize> {
-        self.session.fill()
-    }
-
-    pub fn interest(&self) -> Interest {
-        self.session.interest()
-    }
-
-    pub fn do_handshake(&mut self) -> Result<()> {
-        self.session.do_handshake()
-    }
-
-    pub fn read_buffer_mut(&mut self) -> &mut Buffer {
-        self.session.read_buffer_mut()
-    }
-
-    pub fn write_buffer_mut(&mut self) -> &mut Buffer {
-        self.session.write_buffer_mut()
-    }
-}
-
-impl<Parser, Tx, Rx> Borrow<[u8]> for ClientSession<Parser, Tx, Rx> {
-    fn borrow(&self) -> &[u8] {
-        self.session.borrow()
-    }
-}
-
-impl<Parser, Tx, Rx> Buf for ClientSession<Parser, Tx, Rx> {
-    fn remaining(&self) -> usize {
-        self.session.remaining()
-    }
-
-    fn chunk(&self) -> &[u8] {
-        self.session.chunk()
-    }
-
-    fn advance(&mut self, amt: usize) {
-        self.session.advance(amt)
-    }
-}
-
-unsafe impl<Parser, Tx, Rx> BufMut for ClientSession<Parser, Tx, Rx> {
-    fn remaining_mut(&self) -> usize {
-        self.session.remaining_mut()
-    }
-
-    unsafe fn advance_mut(&mut self, amt: usize) {
-        self.session.advance_mut(amt)
-    }
-
-    fn chunk_mut(&mut self) -> &mut UninitSlice {
-        self.session.chunk_mut()
-    }
-
-    #[allow(unused_mut)]
-    fn put<T: Buf>(&mut self, mut src: T)
-    where
-        Self: Sized,
-    {
-        self.session.put(src)
-    }
-
-    fn put_slice(&mut self, src: &[u8]) {
-        self.session.put_slice(src)
-    }
-}
-
-impl<Parser, Tx, Rx> event::Source for ClientSession<Parser, Tx, Rx> {
-    fn register(&mut self, registry: &Registry, token: Token, interest: Interest) -> Result<()> {
-        self.session.register(registry, token, interest)
-    }
-
-    fn reregister(&mut self, registry: &Registry, token: Token, interest: Interest) -> Result<()> {
-        self.session.reregister(registry, token, interest)
-    }
-
-    fn deregister(&mut self, registry: &Registry) -> Result<()> {
-        self.session.deregister(registry)
-    }
-}
-
-pub struct ServerSession<Parser, Tx, Rx> {
-    session: Session,
-    parser: Parser,
-    pending: VecDeque<Instant>,
-    outstanding: VecDeque<(Option<Instant>, usize)>,
-    _rx: PhantomData<Rx>,
-    _tx: PhantomData<Tx>,
-}
-
-impl<Parser, Tx, Rx> AsRawFd for ServerSession<Parser, Tx, Rx> {
-    fn as_raw_fd(&self) -> i32 {
-        self.session.as_raw_fd()
-    }
-}
-
-impl<Parser, Tx, Rx> Debug for ServerSession<Parser, Tx, Rx> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::result::Result<(), std::fmt::Error> {
-        write!(f, "{:?}", self.session)
-    }
-}
-
-impl<Parser, Tx, Rx> ServerSession<Parser, Tx, Rx>
-where
-    Tx: Compose,
-    Parser: Parse<Rx>,
-{
-    pub fn new(session: Session, parser: Parser) -> Self {
-        Self {
-            session,
-            parser,
-            pending: VecDeque::with_capacity(256),
-            outstanding: VecDeque::with_capacity(256),
-            _rx: PhantomData,
-            _tx: PhantomData,
-        }
-    }
-
-    pub fn into_inner(self) -> Session {
-        self.session
-    }
-
-    pub fn receive(&mut self) -> Result<Rx> {
-        let src: &[u8] = self.session.borrow();
-        match self.parser.parse(src) {
-            Ok(res) => {
-                let now = Instant::now();
-                self.pending.push_back(now);
-                let consumed = res.consumed();
-                let msg = res.into_inner();
-                self.session.consume(consumed);
-                Ok(msg)
-            }
-            Err(e) => Err(e),
-        }
-    }
-
-    pub fn send(&mut self, tx: Tx) -> Result<usize> {
-        SERVER_SESSION_SEND.increment();
-
-        let timestamp = self.pending.pop_front();
-
-        let size = tx.compose(&mut self.session);
-
-        if size == 0 {
-            // we have a zero sized response, increment heatmap now
-            if let Some(timestamp) = timestamp {
-                let now = Instant::now();
-                let latency = now - timestamp;
-                SERVER_RESPONSE_LATENCY.increment(now, latency.as_nanos(), 1);
-            }
-        } else {
-            // we have bytes in our response, we need to add it on the
-            // outstanding response queue
-            self.outstanding.push_back((timestamp, size));
-        }
-
-        Ok(size)
-    }
-
-    pub fn advance_write(&mut self, amt: usize) {
-        if amt == 0 {
-            return;
-        }
-
-        let now = Instant::now();
-
-        let mut amt = amt;
-
-        while amt > 0 {
-            if let Some(mut front) = self.outstanding.pop_front() {
-                if front.1 > amt {
-                    front.1 -= amt;
-                    self.outstanding.push_front(front);
-                    break;
-                } else {
-                    amt -= front.1;
-                    if let Some(ts) = front.0 {
-                        let latency = now - ts;
-                        SERVER_RESPONSE_LATENCY.increment(now, latency.as_nanos(), 1);
-                    }
-                }
-            } else {
-                break;
-            }
-        }
-    }
-
-    /// Attempts to flush all bytes currently in the write buffer to the
-    /// underlying stream. Also handles bookeeping necessary to determine the
-    /// server-side response latency.
-    pub fn flush(&mut self) -> Result<()> {
-        let current_pending = self.session.write_pending();
-        self.session.flush()?;
-        let final_pending = self.session.write_pending();
-
-        let flushed = current_pending - final_pending;
-
-        self.advance_write(flushed);
-
-        Ok(())
-    }
-
-    /// Returns the number of bytes pending in the write buffer.
-    pub fn write_pending(&self) -> usize {
-        self.session.write_pending()
-    }
-
-    /// Reads from the underlying stream and returns the number of bytes read.
-    pub fn fill(&mut self) -> Result<usize> {
-        SERVER_SESSION_READ.increment();
-
-        match self.session.fill() {
-            Ok(amt) => {
-                SERVER_SESSION_READ_BYTES.add(amt as _);
-                Ok(amt)
-            }
-            Err(e) => {
-                if e.kind() != ErrorKind::WouldBlock {
-                    SERVER_SESSION_READ_EX.increment();
-                }
-                Err(e)
-            }
-        }
-    }
-
-    pub fn interest(&self) -> Interest {
-        self.session.interest()
-    }
-
-    pub fn do_handshake(&mut self) -> Result<()> {
-        self.session.do_handshake()
-    }
-
-    pub fn read_buffer_mut(&mut self) -> &mut Buffer {
-        self.session.read_buffer_mut()
-    }
-
-    pub fn write_buffer_mut(&mut self) -> &mut Buffer {
-        self.session.write_buffer_mut()
-    }
-}
-
-impl<Parser, Tx, Rx> Borrow<[u8]> for ServerSession<Parser, Tx, Rx> {
-    fn borrow(&self) -> &[u8] {
-        self.session.borrow()
-    }
-}
-
-impl<Parser, Tx, Rx> Buf for ServerSession<Parser, Tx, Rx> {
-    fn remaining(&self) -> usize {
-        self.session.remaining()
-    }
-
-    fn chunk(&self) -> &[u8] {
-        self.session.chunk()
-    }
-
-    fn advance(&mut self, amt: usize) {
-        self.session.advance(amt)
-    }
-}
-
-unsafe impl<Parser, Tx, Rx> BufMut for ServerSession<Parser, Tx, Rx> {
-    fn remaining_mut(&self) -> usize {
-        self.session.remaining_mut()
-    }
-
-    unsafe fn advance_mut(&mut self, amt: usize) {
-        self.session.advance_mut(amt)
-    }
-
-    fn chunk_mut(&mut self) -> &mut UninitSlice {
-        self.session.chunk_mut()
-    }
-
-    #[allow(unused_mut)]
-    fn put<T: Buf>(&mut self, mut src: T)
-    where
-        Self: Sized,
-    {
-        self.session.put(src)
-    }
-
-    fn put_slice(&mut self, src: &[u8]) {
-        self.session.put_slice(src)
-    }
-}
-
-impl<Parser, Tx, Rx> event::Source for ServerSession<Parser, Tx, Rx> {
-    fn register(&mut self, registry: &Registry, token: Token, interest: Interest) -> Result<()> {
-        self.session.register(registry, token, interest)
-    }
-
-    fn reregister(&mut self, registry: &Registry, token: Token, interest: Interest) -> Result<()> {
-        self.session.reregister(registry, token, interest)
-    }
-
-    fn deregister(&mut self, registry: &Registry) -> Result<()> {
-        self.session.deregister(registry)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        let result = 2 + 2;
-        assert_eq!(result, 4);
     }
 }
